+++ custom_components/linknlink/const.py
@@ -5,5 +5,5 @@ from homeassistant.const import Platform
 DOMAIN = "linknlink"
 CONF_HOST = "host"
 CONF_LOCAL_KEY = "local_key"
-PLATFORMS = (Platform.SENSOR, Platform.BINARY_SENSOR, Platform.EVENT, Platform.SWITCH)
+PLATFORMS = (Platform.SENSOR, Platform.BINARY_SENSOR, Platform.EVENT, Platform.SWITCH, Platform.NUMBER)
 UPDATE_INTERVAL_SECONDS = 30